requirements.txt
NHentai-API==1.0.2
PyQt6==6.3.1